auth_api/python/requirements.txt
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.2
pandas== 1.3.0
protobuf==3.17.3
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
